Deep/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
Pillow>=8.0.0
scipy>=1.7.0
tqdm>=4.60.0
PyYAML>=5.4.0
requests>=2.25.0
